slanker/bot/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
loguru==0.7.2